requirements.txt
fastapi[standard]
uvicorn
pymongo
bcrypt
python-jose
pydantic
python-dotenv
